Add public header pokemon_data.h. WIP.
